requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.2
openpyxl==3.1.2
google-generativeai==0.5.4
python-dotenv
xlsxwriter